Add Thumbnail Images
Inside the Tour container, add three columns of equal width (.col-sm-4).

Inside each column, add an image.

Then, use the .img-thumbnail class to shape the image to a thumbnail.

Normally, you would add the .img-thumbnail class directly to the <img> element. Here, we have placed a thumbnail container around the image, so that we can specify an image text as well.